% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FC_Rationality_Tests.R
\name{bw_rule}
\alias{bw_rule}
\title{Bandwidth selection with rule of thumb}
\usage{
bw_rule(Y, X)
}
\arguments{
\item{Y}{observation}

\item{X}{foreast}
}
\value{
bandwidth as numeric
}
\description{
Bandwidth selection with rule of thumb
}
