% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FC_Rationality_Tests.R
\name{create.instrument.matrix}
\alias{create.instrument.matrix}
\title{Creates instrument matrix from character vector}
\usage{
create.instrument.matrix(X, Y, instruments, other_data = NULL, silent = TRUE)
}
\value{
list containing forecast x, observation y, and instrment marix w
}
\description{
Creates instrument matrix from character vector
}
\examples{
create.instrument.matrix(c(0.1,0.2,0.3),c(1.1,1.2,1.3), c("1","X","lag(Y)"))
}
