% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GDP}
\alias{GDP}
\title{real GDP Greenbook forecasts (1967-2015)}
\format{
A data frame with 176 rows and 2 variables:
\itemize{
  \item date.target: date of the target quarter (1967.2 -- 2015.1)
  \item FC.first, FC.middle, FC.last: Forecasts issued first, closest to the middle, and last in the respective quarter.
  \item Obs.first, Obs.second, Obs.third, Obs.recent: realized GDP growth rate in percentage measured at first, second, third, or most recent vintage
}
}
\source{
\url{https://www.philadelphiafed.org/research-and-data/real-time-center/greenbook-data/philadelphia-data-set}
}
\usage{
GDP
}
\description{
A dataset containing quarterly real GDP growth rates (annualized percentage points) in the United States and according one quarter ahead point forecasts from the Federal Reserve's Greenbook.
}
\details{
Often, multiple forecasts are issued in one quarter,
as the Federal Board meets more than 4 times year
and in irregular time spans.

The forecasts are available as the first and last
forecast issued in any quarter. Additionally,
the column FC.middle denotes the forecast issued
closest to the middle of the quarter.
}
\keyword{datasets}
