% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covest.R
\name{HAC}
\alias{HAC}
\title{HAC estimator}
\usage{
HAC(ids, ...)
}
\arguments{
\item{ids}{values}

\item{...}{}
}
\value{
list of estimation result cov and name (for saving purpose)
}
\description{
HAC estimator
}
