% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convex_combinations.R
\name{id_convex}
\alias{id_convex}
\title{Identification functions of convex combination of mode, mean, and median}
\usage{
id_convex(
  X,
  Y,
  theta = c(1/3, 1/3),
  bw = bw_rule,
  vcov = iid,
  id_mode = id_modegaussian
)
}
\arguments{
\item{X}{foreast}

\item{Y}{observation}

\item{bw}{bandwidth. Numeric or function. (standard selects autamtically with rule of thumb)}
}
\value{

}
\description{
Identification functions of convex combination of mode, mean, and median
}
