% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FC_Rationality_Tests.R
\name{id_mean}
\alias{id_mean}
\title{Identification function of mean}
\usage{
id_mean(Y, X)
}
\arguments{
\item{Y}{observation}

\item{X}{forecast}
}
\description{
Identification function of mean
}
