% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FC_Rationality_Tests.R
\name{id_median}
\alias{id_median}
\title{Identification function of median}
\usage{
id_median(Y, X)
}
\arguments{
\item{Y}{observation}

\item{X}{foreast}
}
\description{
Identification function of median
}
