% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FC_Rationality_Tests.R
\name{id_mode_BiSquare}
\alias{id_mode_BiSquare}
\title{Asymptomatic identification functions of mode based on Bi-square kernel}
\usage{
id_mode_BiSquare(Y, X, bw = bw_rule)
}
\arguments{
\item{Y}{observation}

\item{X}{foreast}

\item{bw}{bandwidth. Numeric or function. (standard selects autamtically with rule of thumb)}
}
\description{
Asymptomatic identification functions of mode based on Bi-square kernel
}
