% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FC_Rationality_Tests.R
\name{id_modegaussian}
\alias{id_modegaussian}
\title{Asymptotic identification functions of mode based on Gaussian kernel}
\usage{
id_modegaussian(Y, X, bw = bw_rule)
}
\arguments{
\item{Y}{observation}

\item{X}{foreast}

\item{bw}{bandwidth. Numeric or function. (standard selects autamtically with rule of thumb)}
}
\description{
Asymptotic identification functions of mode based on Gaussian kernel
}
