% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covest.R
\name{iid}
\alias{iid}
\title{iid variance estimator}
\usage{
iid(ids, ...)
}
\arguments{
\item{ids}{values}

\item{...}{}
}
\value{
list of estimation result cov and name (for saving purpose)
}
\description{
iid variance estimator
}
