% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FC_Rationality_Tests.R
\name{rationality.test}
\alias{rationality.test}
\title{Computes conditional calibration test}
\usage{
rationality.test(
  Y,
  X,
  id.fct = id_mean,
  instruments = NULL,
  vcov = iid,
  bw = NULL,
  ...
)
}
\arguments{
\item{Y}{observation}

\item{X}{forecast}

\item{id.fct}{identification that determines the functional that is tested for}

\item{instruments}{instruments that determine the information tested for}

\item{vcov}{covariance estimation technique. Either a character that is among the predefined methods
or a function that has ids (matrix of identification function values) as input.}

\item{bw}{numeric, function or vector. optional if iden fct requires bandwidth.}
}
\value{

}
\description{
Computes conditional calibration test
}
