% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sce}
\alias{sce}
\title{SCE: Survey of consumer expectations}
\format{
An object of class \code{data.frame} with 4030 rows and 175 columns.
}
\source{
Source: Survey of Consumer Expectations, © 2013-2020 Federal Reserve Bank of New York (FRBNY). The SCE data are available without charge at http://www.newyorkfed.org/microeconomics/sce and may be used subject to license terms posted there. FRBNY disclaims any responsibility for this analysis and interpretation of Survey of Consumer Expectations data.
}
\usage{
sce
}
\description{
Data extracted from the Survey of consumer expectations and the labour market survey of the Federal Reserve Bank of New York.
}
\details{
Expectation is taken from Question oo2e2: "What do you believe your annual earnings will be in 4 months?(dollars per year)"

Income is extracted from Question l3: "How much do you make before taxes and other deductions at your [main/current] job, on an annual basis?
Please include any bonuses, overtime pay, tips or commissions. (dollars per year)"

Labour market survey from https://www.newyorkfed.org/microeconomics/databank.html. Part of the Surevey of Consumer expectations.

Questionaire can be found under https://www.newyorkfed.org/medialibrary/media/research/microeconomics/interactive/downloads/sce-labor-questionnaire.pdf.
}
\keyword{datasets}
