% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convex_combinations.R
\name{test_convex}
\alias{test_convex}
\title{Test convex combinations of mean, median and mode}
\usage{
test_convex(
  data = NULL,
  x = NULL,
  y = NULL,
  instruments = c("1", "X"),
  id_mode = id_modegaussian,
  bw = "bw_rule",
  vcov = iid,
  vcov_inside = iid,
  number_values = 10,
  set_theta = NULL,
  ...
)
}
\arguments{
\item{data}{data input (should contain X and Y)}

\item{x}{alternative input of X directly}

\item{y}{alternative input of Y directly}

\item{instruments}{character describing instruemnts or matrix}

\item{id_mode}{identification function of mode}

\item{bw}{bw selection for mode}

\item{vcov}{covariance estimation for weigting matrix (as function)}

\item{vcov_inside}{covariance estimator in the id_convex function (if NULL mirrors normal covariance estimator in vcov)}

\item{number_values}{number of values for grid on parameter space}

\item{set_theta}{set of parameter values to be tested for. If NULL (standard) creates a set of equi-spaced values with a maximum of number_values in each dimension}

\item{...}{other paramters (use other_data for extracting different instruments)}
}
\description{
Inverts the J-test statistic of the convex combination of mean, median, and mode identification function.
The parts are normalized by a covariance estimator as denoted in vcov_inside. The weighting matrix for the different moments generated by the instrument vector is denoted by vcov.
The standard choice is iid, where the covariance is estimated by its simple sample analogon as proposed in the main paper.
}
